Make function interfaces consistent
ALl functions that take a certificate-type argument (private key, public key,
CSR) now expect a byte representation of the PEM-encoded certificate.

All functions that return a certificate-type argument (private key, public key,
CSR) now return a byte representation of the PEM-encoded certificate.

The fact that the OpenSSL library is being used is now properly hidden within
the functions themselves so that the interface won't need to change if the
library implementation changes

This also adds end to end tests to ensure the functions accept and return the
correct argument types.